+++ HISTORY.rst
@@ -83,6 +83,7 @@ History
      * New descriptor to help set and get from unary relationships like Order.location (rudaporto).
      * Review __raw_acl__ attribute on all models (rudaporto).
      * Improve import to set permissions for each local role imported (rudaporto).
+     * New base class to test dashboard views and test cases for all implemented dashboards: QA, Scout, Professional, Customer (rudaporto).
 
 1.1.0 (2016-10-04)
 ------------------